requirements.txt
boto3==1.19.9
botocore==1.22.9
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
docopt==0.6.2
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
json2html==1.3.0
MarkupSafe==2.0.1
pipreqs==0.4.11
python-dateutil==2.8.2
python-dotenv==0.19.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
yarg==0.1.9
